Update 782 - v11

+++ Libraries/ReikaAABBHelper.java
@@ -142,11 +142,11 @@ public final class ReikaAABBHelper extends DragonAPICore {
 	}
 
 	/** Returns a sized bounding box centered on a Blocks. Args: x, y, z */
-	public static AxisAlignedBB getBlockCenteredAABB(int x, int y, int z, int range) {
+	public static AxisAlignedBB getBlockCenteredAABB(int x, int y, int z, double range) {
 		return AxisAlignedBB.getBoundingBox(x, y, z, x+1, y+1, z+1).expand(range, range, range);
 	}
 
-	public static AxisAlignedBB getEntityCenteredAABB(Entity e, int range) {
+	public static AxisAlignedBB getEntityCenteredAABB(Entity e, double range) {
 		return AxisAlignedBB.getBoundingBox(e.posX, e.posY, e.posZ, e.posX, e.posY, e.posZ).expand(range, range, range);
 	}
